Create Pipeline_7_Multiple_Loops.py
Here is the Pipeline that supports C Programs with Multiple Loops i.e. 2 loops for example. 

LLM generates the invariants for the Loop(s) in the C program that have the respective Loop Heads labelled with comments // Line A ... // Line B etc. 

And then if the generated invariants are syntactically correct and they match the regex then they get inserted into the c program in the appropriate positions.

Note: More experimentation with prompts are needed in this pipeline.